requirements.txt
colorama
python-dotenv
groq
colorama